Update Definition of Done Team <team name>.md
Definition of done for team <team name>. Determines if a user story on the sprint backlog is done.

Please check for completeness and spelling errors. Remember that if a user story requires less or more stringent exit criteria, we can agree, with approval from the Product Owner, to deviate from this document.